backend/requirements.txt
fastapi
uvicorn
python-multipart
faster-whisper
gspread
oauth2client
moviepy
python-dotenv
pandas
static-ffmpeg
